Feat: Add function to snippet tool buttons
