+++ src/components/Snippet/SnippetTool/SnippetTool.js
@@ -1,8 +1,21 @@
+import { useState } from "react";
+import { useHistory } from "react-router";
 import styled from "styled-components";
 
 import Button from "../../common/Button";
+import SnippetSavingModal from "../../SnippetSavingModal";
+
+import { deleteSnippet } from "../../../api/service";
 
 import { TOOL } from "../../../constants/variants";
+import { OK } from "../../../constants/messages";
+
+import {
+  COPY,
+  SAVE,
+  SHARE,
+  DELETE
+} from "../../../constants/names";
 
 const ToolWrapper = styled.div`
   display: flex;
@@ -10,10 +23,60 @@ const ToolWrapper = styled.div`
   width: 320px;
 `;
 
-export default function SnippetTool() {
+export default function SnippetTool({ creator, language, code, snippetId }) {
+  const history = useHistory();
+  const [isOpened, setIsOpened] = useState(false);
+
+  const handleClick = async (event) => {
+    const buttonName = event.target.textContent;
+
+    if (buttonName === SAVE) {
+      setIsOpened(true);
+      return;
+    }
+
+    if (buttonName === COPY) {
+      await navigator.clipboard.writeText(code);
+
+      alert("copied!");
+      return;
+    }
+
+    if (buttonName === SHARE) {
+
+      return;
+    }
+
+    if (buttonName === DELETE) {
+      const { result } = await deleteSnippet({ id: snippetId });
+
+      if (result === OK) {
+        history.push("/");
+      }
+    }
+  };
+
   return (
-    <ToolWrapper>
-      {["저장", "복사", "공유", "삭제"].map((name) => <Button key={name} variant={TOOL}>{name}</Button>)}
-    </ToolWrapper>
+    <>
+      <ToolWrapper>
+        {[SAVE, COPY, SHARE, DELETE].map((name) => (
+          <Button
+            key={name}
+            variant={TOOL}
+            onClick={(event) => handleClick(event)}
+          >
+            {name}
+          </Button>
+        ))}
+      </ToolWrapper>
+      {isOpened && (
+        <SnippetSavingModal
+          creator={creator}
+          language={language}
+          code={code}
+          onClick={setIsOpened}
+        />
+      )}
+    </>
   );
 }
