+++ src/components/SnippetDetailPage/DetailSnippet.js
@@ -5,6 +5,7 @@ import UserProfile from "../Snippet/UserProfile/UserProfile";
 import SnippetTool from "../Snippet/SnippetTool/SnippetTool";
 import SnippetInfo from "../Snippet/SnippetInfo/SnippetInfo";
 import CodeEditor from "../CodeEditor/CodeEditor";
+import { useState } from "react";
 
 const InfoWrapper = styled.div`
   display: flex;
@@ -30,8 +31,10 @@ const Date = styled.div`
 `;
 
 export default function DetailSnippet({ snippet }) {
-  const { hashtagList, poster, language, likerList, commentList, createdAt, code } = snippet;
+  const { hashtagList, creator, poster, language, likerList, commentList, createdAt, _id, code: defaultCode } = snippet;
   const { nickname, imageUrl, followerList } = poster;
+
+  const [code, setCode] = useState(defaultCode);
 
   const userId = localStorage.getItem("userId");
   const isLiked = likerList.indexOf(userId) !== -1;
@@ -43,7 +46,7 @@ export default function DetailSnippet({ snippet }) {
       <HashtagList
         type="detail"
         hashtags={hashtagList} />
-      <CodeEditor width="1100px" height="400px" code={code} />
+      <CodeEditor width="1100px" height="400px" code={code} onEdit={setCode} />
       <InfoWrapper>
         <UserProfile
           profileUrl={imageUrl}
@@ -60,7 +63,12 @@ export default function DetailSnippet({ snippet }) {
       </InfoWrapper>
       <Footer>
         <Date>{formatDate}</Date>
-        <SnippetTool />
+        <SnippetTool
+          creator={creator._id}
+          language={language}
+          code={code}
+          snippetId={_id}
+        />
       </Footer>
     </>
   );
