+++ src/constants/names.js
@@ -1,9 +1,15 @@
 const REGISTRATION = "등록";
 const CANCEL = "취소";
 const DELETE = "삭제";
+const SAVE = "저장";
+const COPY = "복사";
+const SHARE = "공유";
 
 export {
   REGISTRATION,
   CANCEL,
   DELETE,
+  SAVE,
+  COPY,
+  SHARE,
 };
